requirements.txt
anyio==3.7.1
async-timeout==4.0.3
certifi==2023.7.22
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.7.0
psycopg2-binary==2.9.9
python-telegram-bot==20.3
redis==5.0.1
sniffio==1.3.0
typing_extensions==4.7.1
zipp==3.15.0
